modify itempoint
1. block define as uint64 for fileid
2. offset define as uint32 for file offset
3. fix all elog itempoint

+++ src/backend/access/aocs/aocsam.c
@@ -328,7 +328,7 @@ initscan_with_colinfo(AOCSScanDesc scan)
 	scan->cur_seg = -1;
 	scan->cur_seg_row = 0;
 
-	ItemPointerSet(&scan->cdb_fake_ctid, 0, 0);
+	ItemPointerSet(&scan->cdb_fake_ctid, (uint64)0, 0);
 
 	scan->totalBytesRead = 0;
